Added a progress bar
