+++ CommandScanner/CommandScanner.csproj
@@ -59,11 +59,7 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Page Include="Windows\LoadScreen.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Windows\MainWindow.xaml">
+    <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -73,11 +69,8 @@
     </Compile>
     <Compile Include="HelperClasses\ConnectionService.cs" />
     <Compile Include="HelperClasses\Enums.cs" />
-    <Compile Include="SystemCommands\Command.cs" />
-    <Compile Include="Windows\LoadScreen.xaml.cs">
-      <DependentUpon>LoadScreen.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Windows\MainWindow.xaml.cs">
+    <Compile Include="HelperClasses\Command.cs" />
+    <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
