--- CommandScanner/Windows/LoadScreen.xaml.cs
@@ -1,20 +1,0 @@
-﻿using System.Windows;
-
-namespace CommandScanner
-{
-	/// <summary>
-	/// Interaction logic for LoadScreen.xaml
-	/// </summary>
-	public partial class LoadScreen
-	{
-		public LoadScreen()
-		{
-			InitializeComponent();
-		}
-
-		private void scanningProgress_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
-		{
-
-		}
-	}
-}
